pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chirho"
version = "0.3.0"
description = "Causal reasoning"
readme = "README.rst"
requires-python = ">=3.9"
license = {text = "Apache 2.0"}
authors = [{name = "Basis"}]
keywords = [
  "machine learning",
  "statistics",
  "probabilistic programming",
  "bayesian modeling",
  "pytorch"
]
classifiers = [
  "Intended Audience :: Developers",
  "Intended Audience :: Education",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: POSIX :: Linux",
  "Operating System :: MacOS :: MacOS X",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
]
dependencies = [
  "pyro-ppl>=1.9.1",
]

[project.urls]
Homepage = "https://www.basis.ai/"
Source = "https://github.com/BasisResearch/chirho"

[project.optional-dependencies]
dynamical = [
  "torchdiffeq",
]
extras = [
  "jupyter",
  "graphviz",
  "matplotlib",
  "pandas",
  "seaborn",
  "pytorch-lightning",
  "scikit-image",
  "tensorboard",
  "typing_extensions",
]
test = [
  "chirho[dynamical,extras]",
  "typing_extensions",
  "pytest",
  "pytest-cov",
  "pytest-xdist",
  "mypy==1.10",
  "ruff",
  "sphinx==7.1.2",
  "sphinxcontrib-bibtex",
  "sphinxcontrib-jquery",
  "sphinx_rtd_theme==1.3.0",
  "myst_parser",
  "nbsphinx",
  "nbval",
  "nbqa",
]

[dependency-groups]
dev = ["chirho[dynamical,extras,test]"]

[tool.setuptools.packages.find]
include = ["chirho", "chirho.*"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
  "F",     # pyflakes
  "I",     # isort
  "PERF",  # perflint
  "UP",    # pyupgrade
]
ignore = [
  "UP031",   # Use format specifiers instead of percent format
  "PERF401", # Use a list comprehension to create a transformed list
]
